\name{human.protein.names}
\alias{human.protein.names}
\alias{my.protein.info}
\title{
    Info on proteins
}
\description{
    Gather human readable information from protein group codes.
}
\usage{

my.protein.info(x, protein.g)


human.protein.names(my.protein.info)
}
\arguments{
  \item{x}{ProteinGroup object}
  \item{protein.g}{protein}
  \item{my.protein.info}{Return value of function my.protein.info}
}
\author{Florian P Breitwieser}
