\name{isobar-package}
\alias{isobar-package}
\alias{isobar}
\docType{package}
\title{
Analysis and quantitation of isobaric tag Proteomics data
}
\description{
isobar provides methods for preprocessing, normalization, and
report generation for the analysis of Proteomics data generated by Mass
spectrometers with the usage of isobaric tags, such as iTRAQ and TMT.
}
\details{
\tabular{ll}{
Package: \tab isobar\cr
Version: \tab 0.1\cr
Depends: \tab R (>= 2.9.0), methods, Biobase, ggplot2, Hmisc, RColorBrewer, distr, geneplotter\cr
LazyLoad: \tab yes\cr
License: \tab GPL-2\cr
URL: \tab http://bioinformatics.cemm.oeaw.ac.at\cr
Collate: \tab utils.R
mgf-utils.R
ProteinGroup-class.R
IBSpectra-class.R
NoiseModel-class.R
ratio-methods.R
zzz.R\cr
Built: \tab R 2.11.1; x86_64-unknown-linux-gnu; 2010-09-14 20:08:47 UTC; unix\cr
}

Index:
\preformatted{
IBSpectra-class         IBSpectra objects
NoiseModel-class        NoiseModel objects
ProteinGroup-class      ProteinGroup objects
isobar-analysis         IBSpectra analysis: Protein ratios and
                        distribution fitting
isobar-import           IBSpectra Import
isobar-plots            IBSpectra plots
isobar-preprocessing    IBSpectra preprocessing
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{isobar-analyzing-experiments} \tab Isobar - Package structure (source, pdf)\cr
\code{isobar-pacakge-structure} \tab Isobar - Package structure (source, pdf)\cr
\code{isobar} \tab Isobar Overview (source, pdf)\cr
}
}
\author{
Florian P. Breitwieser <fbreitwieser@cemm.oeaw.ac.at>,
Jacques Colinge <jcolinge@cemm.oeaw.ac.at>

Maintainer: Florian P. Breitwieser <fbreitwieser@cemm.oeaw.ac.at>
}
\keyword{ package }