\name{reporter.protein-methods}
\docType{methods}
\alias{reporter.protein-methods}
\alias{reporter.protein}
\alias{reporter.protein,ProteinGroup,character-method}
\title{Get reporter protein group identifier for protein group identifier}
\description{
 Methods for function \code{reporter.protein} in package \pkg{isobar}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ProteinGroup", protein.g = "character")}}{
 Get reporter protein for protein group identifier.
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
