\name{shared.ratios.sign}
\alias{shared.ratios.sign}
\title{
  Plot and get significantly shared ratios.
}
\description{
  Plot and get significantly shared ratios.
}
\usage{
shared.ratios.sign(ress, z.shared, min.spectra = 1, plot = TRUE)
}
\arguments{
  \item{ress}{Result of shared.ratios.}
  \item{z.shared}{z. }
  \item{min.spectra}{Minimal number of spectra needed.}
  \item{plot}{plot. }
}
\author{
  Florian P.\ Breitwieser
}

\seealso{
  \code{\link{shared.ratios}}.
}

