\name{specificities}
\alias{specificities}
\alias{group-specific}
\alias{reporter-specific}
\alias{unspecific}
\alias{UNSPECIFIC}
\alias{GROUPSPECIFIC}
\alias{REPORTERSPECIFIC}
\alias{SPECIFICITIES}
\docType{data}
\title{
    Peptide specificities
}
\description{
    Peptides can appear in multiple proteins and therefore have different specificities.
}
\details{
    reporter specific: peptides specific to reporter.
    group specific: peptides specific to the group.
    unspecific: peptides shared with other proteins.
}
\keyword{datasets}
