\name{Isobar util functions}
\alias{paste0}
\title{
  Isobar util functions
}
\description{
  Utility functions.
  paste0 as a shorthand to paste(...,sep="") in versions of R pre 2.14.
  
}
\usage{
paste0(..., sep = "")
}
\arguments{
  \item{\dots}{Arguments to paste.}
  \item{sep}{Separator.}
}
\author{
  Florian P Breitwieser
}
