\name{writeHscoreData}
\alias{writeHscoreData}
\title{
  Write identifications into a format suitable for Hscore.
}
\description{
  Write identifications into a format suitable for Hscore.
}
\usage{
writeHscoreData(outfile, ids, massfile = "defs.txt")
}
\arguments{
  \item{outfile}{Output file. }
  \item{ids}{ IBSpectra identifications data.frame (ie fData). } 
  \item{massfile}{ Definition file for Hscore. }
}
\author{Florian P. Breitwieser }

