\name{ratiosReshapeWide}
\alias{ratiosReshapeWide}

\title{Reshape output of proteinRatios into wide format}
\description{
  Reshape output of proteinRatios into wide format
}

\usage{
ratiosReshapeWide(quant.tbl, grouped.cols = TRUE)
}
\arguments{
  \item{quant.tbl}{Output of proteinRatios or peptideRatios.}
  \item{grouped.cols}{Whether the columns should be grouped next to each other.}
}
\author{Florian P. Breitwieser}
